// src/app/types/profile-response.type.ts
// export interface ProfileResponseType {
//     profiles: Profile[];
// }

export interface ProfileInt {
    [x: string]: any;
    evalassigned: {};
    status: {};
    experience: number;
    id: number;
    location: string;
    mobileno: number;
    name: string;
    skill1: string;
    skill2: string;
    vendor: string;
}